% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotSkittles}
\alias{plotSkittles}
\title{Plot the Skittles color palettes}
\usage{
plotSkittles()
}
\value{
None.
}
\description{
Plot illustrating the Skittles and M&M's color palettes
}
\examples{
plotSkittles()
}
\seealso{
\code{\link{RSkittleBrewer}}, \code{\link{plotSmarties}}
}

