% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSmarties.R
\name{plotSmarties}
\alias{plotSmarties}
\title{Plot the Smarties color palettes}
\usage{
plotSmarties()
}
\value{
None.
}
\description{
Plot illustrating the Smarties color palette
}
\examples{
plotSmarties()
}
\seealso{
\code{\link{RSkittleBrewer}}, \code{\link{plotSkittles}}
}

