% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RSkittleBrewer}
\alias{RSkittleBrewer}
\title{Candy-based color palettes}
\usage{
RSkittleBrewer(flavor = c("original", "tropical", "wildberry", "M&M",
  "smarties"))
}
\arguments{
  \item{flavor}{Character string for candy-based color
  palette.}
}
\value{
Vector of character strings representing the chosen set of colors.
}
\description{
Vectors of colors corresponding to different candies.
}
\examples{
plotSkittles()
plotSmarties()
}
\seealso{
\code{\link{plotSkittles}}, \code{\link{plotSmarties}}
}
\keyword{hplot}

