% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSkittles.R
\name{plotSkittles}
\alias{plotSkittles}
\title{Plot the Skittles color palettes}
\usage{
plotSkittles()
}
\value{
None.
}
\description{
Plot illustrating the Skittles and M&M's color palettes
}
\examples{
plotSkittles()
}
\seealso{
\code{\link{RSkittleBrewer}}, \code{\link{plotSmarties}}
}

