% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotSmarties}
\alias{plotSmarties}
\title{Plot the Smarties color palettes}
\usage{
plotSmarties()
}
\value{
None.
}
\description{
Plot illustrating the Smarties color palette
}
\examples{
plotSmarties()
}
\seealso{
\code{\link{RSkittleBrewer}}, \code{\link{plotSkittles}}
}

